refactor: #21 패키지 수정
